Integrate test run summary
