+++ web/src/pages/Home/Home.tsx
@@ -1,10 +1,9 @@
 import Layout from 'components/Layout';
 import withAnalytics from 'components/WithAnalytics/WithAnalytics';
-import ExperimentalFeature from 'utils/ExperimentalFeature';
 import Content from './Content';
 
 const Home = () => (
-  <Layout hasMenu={ExperimentalFeature.isEnabled('transactions')}>
+  <Layout hasMenu>
     <Content />
   </Layout>
 );
