+++ web/src/pages/Home/TestList.tsx
@@ -2,15 +2,13 @@ import {useCallback} from 'react';
 import {useNavigate} from 'react-router-dom';
 
 import InfiniteScroll from 'components/InfiniteScroll';
-import TestCard from 'components/TestCard';
-import TestCardV2 from 'components/TestCard/TestCardV2';
+import TestCard from 'components/TestCard/TestCard';
 import useDeleteTest from 'hooks/useDeleteTest';
 import useInfiniteScroll from 'hooks/useInfiniteScroll';
 import useTestCrud from 'providers/Test/hooks/useTestCrud';
 import {useGetTestListQuery} from 'redux/apis/TraceTest.api';
 import HomeAnalyticsService from 'services/Analytics/HomeAnalytics.service';
 import {TTest} from 'types/Test.types';
-import ExperimentalFeature from 'utils/ExperimentalFeature';
 import * as S from './Home.styled';
 import NoResults from './NoResults';
 
@@ -50,13 +48,9 @@ const TestList = ({query}: IProps) => {
       emptyComponent={<NoResults />}
     >
       <S.TestListContainer data-cy="test-list">
-        {list?.map(test =>
-          ExperimentalFeature.isEnabled('transactions') ? (
-            <TestCardV2 key={test.id} onDelete={onDelete} onRun={onRunTest} onViewAll={onClick} test={test} />
-          ) : (
-            <TestCard test={test} onClick={onClick} onDelete={onDelete} onRunTest={onRunTest} key={test.id} />
-          )
-        )}
+        {list?.map(test => (
+          <TestCard key={test.id} onDelete={onDelete} onRun={onRunTest} onViewAll={onClick} test={test} />
+        ))}
       </S.TestListContainer>
     </InfiniteScroll>
   );
