+++ web/src/pages/Test/Test.tsx
@@ -3,14 +3,13 @@ import {useParams} from 'react-router-dom';
 import Layout from 'components/Layout';
 import withAnalytics from 'components/WithAnalytics/WithAnalytics';
 import TestProvider from 'providers/Test';
-import ExperimentalFeature from 'utils/ExperimentalFeature';
 import TestContent from './TestContent';
 
 const Test = () => {
   const {testId = ''} = useParams();
 
   return (
-    <Layout hasMenu={ExperimentalFeature.isEnabled('transactions')}>
+    <Layout hasMenu>
       <TestProvider testId={testId}>
         <TestContent />
       </TestProvider>
